requirements.txt
streamlit
duckdb
pandas
plotly
faker
python-dotenv
slack_sdk
matplotlib
